define a public testvectors crate. Necessary for zcash-local-net port, has additional benefits
