+++ libtonode-tests/tests/sync.rs
@@ -1,4 +1,5 @@
 use tempfile::TempDir;
+use testvectors::seeds::HOSPITAL_MUSEUM_SEED;
 use zingo_netutils::GrpcConnector;
 use zingo_sync::sync::sync;
 use zingolib::{
@@ -6,7 +7,6 @@ use zingolib::{
     get_base_address_macro,
     lightclient::LightClient,
     testutils::{lightclient::from_inputs, scenarios},
-    testvectors::seeds::HOSPITAL_MUSEUM_SEED,
     wallet::WalletBase,
 };
