+++ zingolib/src/lightclient.rs
@@ -707,7 +707,7 @@ async fn get_recent_median_price_from_gemini() -> Result<f64, PriceFetchError> {
 #[cfg(test)]
 mod tests {
     use crate::config::{ChainType, RegtestNetwork, ZingoConfig};
-    use crate::testvectors::seeds::CHIMNEY_BETTER_SEED;
+    use testvectors::seeds::CHIMNEY_BETTER_SEED;
     use tokio::runtime::Runtime;
 
     use crate::{lightclient::LightClient, wallet::WalletBase};
