+++ zingolib/src/mocks.rs
@@ -627,7 +627,7 @@ pub mod proposal {
             let mut builder = Self::new();
             builder
                 .recipient_address(
-                    address_from_str(crate::testvectors::REG_O_ADDR_FROM_ABANDONART).unwrap(),
+                    address_from_str(testvectors::REG_O_ADDR_FROM_ABANDONART).unwrap(),
                 )
                 .amount(NonNegativeAmount::from_u64(100_000).unwrap());
             builder
